moves sectionheader component to utilities and updates preTitle prop
