+++ resources/assets/components/Quiz/QuizQuestion.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import QuizChoice from './QuizChoice';
-import SectionHeader from '../SectionHeader';
+import SectionHeader from '../utilities/SectionHeader/SectionHeader';
 import { convertNumberToWord } from '../../helpers';
 
 const QuizQuestion = props => {
@@ -26,13 +26,13 @@ const QuizQuestion = props => {
     />
   ));
 
-  const preTitle = hideQuestionNumber
+  const superTitle = hideQuestionNumber
     ? null
     : `Question ${convertNumberToWord(Number(id) + 1)}`;
 
   return (
     <div className="question">
-      <SectionHeader preTitle={preTitle} title={title} />
+      <SectionHeader superTitle={superTitle} title={title} />
       <div className="question__choices">{quizChoices}</div>
     </div>
   );
