--- resources/assets/components/SectionHeader/index.js
@@ -1,1 +1,0 @@
-export default from './SectionHeader';
