+++ resources/assets/components/utilities/SectionHeader/section-header.scss
@@ -1,4 +1,4 @@
-@import '../../scss/next-toolbox.scss';
+@import '../../../scss/next-toolbox.scss';
 
 .section-header {
   width: 100%;
